refactor(agent_meta): extract secret-redaction helpers to _meta/secrets.py
Pure move: _SECRET_PATTERNS and _redact_secrets are now in
_state/_meta/secrets.py and re-exported from agent_meta to preserve
the public attribute surface (am._redact_secrets etc.). No behavior
change; 412 _state/ tests pass.
